Rework entire container
Added creation of a user and a script to wrap bitbake to drop root
privileges before building since bitbake really does not like having
root privileges. Additionally moved bits around in the Dockerfile to
prevent recreation of layers that likely won't change much. Switched
from using CMD to ENTRYPOINT because we'll always be kicking off the
container with additional commands and we want the my_init script to run
beforehand.